Updated logic with LBFGS
Internal testing found a boundary condition where setting the step size
initially takes an exceedingly long time.  This logic updates that so a
more approximate step size is found by the solver in this situation.
This is fine, as there is a subsequent step to ensure that the resulting
step is not so large as to increase the function value, making the model
robust against poorly chosen step sizes.